refactor: split calldata into modules, cleanup, tests fix response, cairo test echo merge
