+++ src/utils/calldata/index.ts
@@ -1,0 +1,161 @@
+import BN from 'bn.js';
+import assert from 'minimalistic-assert';
+
+import { Abi, AbiEntry, Args, Calldata, FunctionAbi, abiStructs } from '../../types';
+import { felt, isLen } from './cairo';
+import formatter from './formatter';
+import { parseCalldataField } from './requestParser';
+import responseParser from './responseParser';
+import validateFields from './validate';
+
+// Helpers
+
+export class CallData {
+  abi: Abi;
+
+  protected readonly structs: abiStructs;
+
+  constructor(abi: Abi) {
+    this.abi = abi;
+    this.structs = CallData.getAbiStruct(abi);
+  }
+
+  /**
+   * Validates if all arguments that are passed to the method are corresponding to the ones in the abi
+   *
+   * @param type - type of the method
+   * @param method  - name of the method
+   * @param args - arguments that are passed to the method
+   */
+  public validate(type: 'INVOKE' | 'CALL' | 'DEPLOY', method: string, args: Array<any> = []) {
+    // ensure provided method of type exists
+    if (type !== 'DEPLOY') {
+      const invocableFunctionNames = this.abi
+        .filter((abi) => {
+          if (abi.type !== 'function') return false;
+          const isView = abi.stateMutability === 'view';
+          return type === 'INVOKE' ? !isView : isView;
+        })
+        .map((abi) => abi.name);
+      assert(
+        invocableFunctionNames.includes(method),
+        `${type === 'INVOKE' ? 'invocable' : 'viewable'} method not found in abi`
+      );
+    }
+
+    // get requested method from abi
+    const abiMethod = this.abi.find((abi) =>
+      type === 'DEPLOY'
+        ? abi.name === method && abi.type === method
+        : abi.name === method && abi.type === 'function'
+    ) as FunctionAbi;
+
+    // validate arguments length
+    const inputsLength = CallData.abiInputsLength(abiMethod.inputs);
+    if (args.length !== inputsLength) {
+      throw Error(
+        `Invalid number of arguments, expected ${inputsLength} arguments, but got ${args.length}`
+      );
+    }
+
+    // validate parameters
+    validateFields(abiMethod, args, this.structs);
+  }
+
+  /**
+   * Parse the calldata by using input fields from the abi for that method
+   *
+   * @param args - arguments passed the the method
+   * @param inputs  - list of inputs(fields) that are in the abi
+   * @return {Calldata} - parsed arguments in format that contract is expecting
+   */
+  public compile(args: Array<any>, inputs: AbiEntry[]): Calldata {
+    const argsIterator = args[Symbol.iterator]();
+    return inputs.reduce(
+      (acc, input) =>
+        isLen(input.name) ? acc : acc.concat(parseCalldataField(argsIterator, input, this.structs)),
+      [] as Calldata
+    );
+  }
+
+  /**
+   * Compile contract callData without abi
+   * @param data Object representing cairo method arguments
+   * @returns string[]
+   */
+  static compile(data: object): string[] {
+    const createTree = (obj: object) => {
+      const getEntries = (o: object, prefix = ''): any => {
+        const oe = Array.isArray(o) ? [o.length.toString(), ...o] : o;
+        return Object.entries(oe).flatMap(([k, v]) => {
+          const kk = Array.isArray(oe) && k === '0' ? '$$len' : k;
+          if (BN.isBN(v)) return [[`${prefix}${kk}`, felt(v)]];
+          return Object(v) === v ? getEntries(v, `${prefix}${kk}.`) : [[`${prefix}${kk}`, felt(v)]];
+        });
+      };
+      return Object.fromEntries(getEntries(obj));
+    };
+
+    // flatten structs, tuples, add array length. Process leafs as Felt
+    const callTree = createTree(data);
+    // convert to array
+    const callTreeArray = Object.values(callTree);
+
+    // add compiled property to array object
+    Object.defineProperty(callTreeArray, 'compiled', {
+      enumerable: false,
+      writable: false,
+      value: true,
+    });
+    return callTreeArray;
+  }
+
+  /**
+   * Parse elements of the response array and structuring them into response object
+   * @param method - method name
+   * @param response  - response from the method
+   * @return - parsed response corresponding to the abi
+   */
+  public parse(method: string, response: string[]): Object {
+    const { outputs } = this.abi.find((abi) => abi.name === method) as FunctionAbi;
+    const responseIterator = response.flat()[Symbol.iterator]();
+
+    return outputs.flat().reduce((acc, output) => {
+      acc[output.name] = responseParser(responseIterator, output, this.structs, acc);
+      if (acc[output.name] && acc[`${output.name}_len`]) {
+        delete acc[`${output.name}_len`];
+      }
+      return acc;
+    }, {} as Args);
+  }
+
+  /**
+   * Format cairo method response data to native js values based on provided format schema
+   * @param method - cairo method name
+   * @param response - cairo method response
+   * @param format - formatter object schema
+   * @returns parsed and formatted response object
+   */
+  public format(method: string, response: string[], format: Object): Object {
+    const parsed = this.parse(method, response);
+    return formatter(parsed, format);
+  }
+
+  // Helper to calculate inputs
+  static abiInputsLength(inputs: AbiEntry[]) {
+    return inputs.reduce((acc, input) => (!isLen(input.name) ? acc + 1 : acc), 0);
+  }
+
+  // Helper to extract structs
+  static getAbiStruct(abi: Abi) {
+    return abi
+      .filter((abiEntry) => abiEntry.type === 'struct')
+      .reduce(
+        (acc, abiEntry) => ({
+          ...acc,
+          [abiEntry.name]: abiEntry,
+        }),
+        {}
+      );
+  }
+}
